moved quote handling to dedicated class
